+++ Core/DolphinVM/finalize.cpp
@@ -41,7 +41,7 @@ void ObjectMemory::finalize(OTE* ote)
 	#endif
 		
 	MemoryManager* memMan = memoryManager();
-	Interpreter::queueForFinalization(ote, (unsigned)integerValueOf(memMan->m_hospiceHighWater));
+	Interpreter::queueForFinalization(ote, integerValueOf(memMan->m_hospiceHighWater));
 }
 
 void Interpreter::scheduleFinalization()
@@ -80,9 +80,9 @@ void Interpreter::scheduleFinalization()
 	}
 
 	MemoryManager* memMan = ObjectMemory::memoryManager();
-	unsigned count = m_qForFinalize.Count();
+	size_t count = m_qForFinalize.Count();
 	// Raise interrupt when at or above high water mark
-	if (count >= (unsigned)integerValueOf(memMan->m_hospiceHighWater))
+	if (count >= static_cast<size_t>(integerValueOf(memMan->m_hospiceHighWater)))
 		queueInterrupt(VMI_HOSPICECRISIS, integerObjectOf(count));
 }
